pyproject.toml
[project]
name = "system-tables"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "databricks-sdk>=0.73.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]
